Snake-Classification/requirements.txt
pillow==11.0.0
PyYAML==6.0.2
tensorboard==2.19.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchmetrics==1.7.1
torchvision==0.21.0+cu126
tqdm==4.67.1
